fix: 🐛 video stuttering
When video elements are not visible in current viewport, Chromium lower
the priority of video rendering for the performance sake. This results
in stuttering videos. To mitigate this we need to put video elements in
the viewport, so I put it at the top-left of the window but with low
opacity and mimimum size